old-tests/submission/W3C/domhtml/HTMLAppletElement11.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Applet</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-applet-object" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<applet align="bottom" alt="Applet Number 1" archive="" codebase="applets" height="306" hspace="0" name="applet1" object="./DOMTSApplet.dat" vspace="0" width="301"></applet>
</p>


<script>
var domtest = async_test("The object IDL attribute must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vobject;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("applet");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vobject = testNode.object;

      assert_uri_equals(document.location.protocol,null,document.location.host,"DOMTSApplet.dat",null,null,null,true,vobject);
       
    domtest.done();
}));

</script>
<p>The object IDL attribute must reflect the respective content attribute of the same name. For the purposes of reflection, the applet element's object content attribute is defined as containing a URL.</p>
<div id="log"></div>
</body></html>
